labs

+++ lab04/src/task2.c
@@ -3,9 +3,9 @@ int main()
 	#define NUM 123.987 //Задаємо число
 
 	int part1 = NUM; //Виводимо цілу частину
-	float part2 = (NUM - part1) * 1000; //Виводимо дробову частину
+	int part2 = (NUM - part1) * 1000; //Виводимо дробову частину
 
-	float diff = (part2 / part1); //Знаходимо частку
+	float diff = ((float)part2 / part1); //Знаходимо частку
 	diff *= 100; //Зводимо частку до 2 знаків після коми наступними командами
 	diff = (int)diff;
 	diff /= 100;
